Move stuff to ci folder
