--- .jenkins/README.md
@@ -1,3 +1,0 @@
-Jenkins CI
-
-Please run everything in Docker!
