+++ ci/benchmarks/.gitignore
@@ -1,0 +1,2 @@
+*.csv
+*.png
