+++ ci/jenkins/README.md
@@ -1,0 +1,3 @@
+Jenkins CI
+
+Please run everything in Docker!
